requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
google-genai==1.67.0
requests==2.31.0
Werkzeug==2.3.7
